refactor: set user and agency as same entity

+++ src/domains/users/services/UpdateAvatarService.ts
@@ -6,16 +6,13 @@ import IUsersRepository from '@domains/users/rules/IUsersRepository';
 import AppError from '@shared/errors/AppError';
 import IStorageProvider from '@shared/providers/StorageProvider/rules/IStorageProvider';
 import User from '../infra/typeorm/entities/User';
-import IAgencyRepository from '../rules/IAgencyRepository';
-import Agency from '../infra/typeorm/entities/Agency';
-import { UserTypes } from '../enums/UserEnums';
 
 interface IRequest {
   user_id: string;
   avatarFilename: string;
 }
 interface IResponse {
-  user: User | Agency;
+  user: User;
   user_type: number;
 }
 
@@ -25,9 +22,6 @@ class UpdateUserAvatarService {
     @inject('UsersRepository')
     private usersRepository: IUsersRepository,
 
-    @inject('AgencyRepository')
-    private agencyRepository: IAgencyRepository,
-
     @inject('StorageProvider')
     private storageProvider: IStorageProvider,
   ) {}
@@ -36,37 +30,19 @@ class UpdateUserAvatarService {
     user_id,
     avatarFilename,
   }: IRequest): Promise<IResponse> {
-    let user: User | Agency | undefined;
-    let user_type = 0;
+    const user_type = 0;
 
-    user = await this.usersRepository.findById(user_id);
+    const user = await this.usersRepository.findById(user_id);
 
-    if (!user) {
-      user = await this.agencyRepository.findById(user_id);
-      if (!user) {
-        throw new AppError('User not found');
-      } else {
-        user_type = UserTypes.EnvironmentalAgency;
-      }
-    } else {
-      user_type = UserTypes.Reporter;
-    }
+    if (!user) throw new AppError('User not found');
 
-    if (user.avatar) {
-      await this.storageProvider.deleteFile(user.avatar);
-    }
+    if (user.avatar) await this.storageProvider.deleteFile(user.avatar);
 
     const filename = await this.storageProvider.saveFile(avatarFilename);
 
     user.avatar = filename;
 
-    if (user_type === UserTypes.Reporter) {
-      await this.usersRepository.update(user as User);
-    }
-
-    if (user_type === UserTypes.EnvironmentalAgency) {
-      await this.agencyRepository.update(user as Agency);
-    }
+    await this.usersRepository.update(user);
 
     return { user: classToClass(user), user_type };
   }
